+++ android/app/src/main/java/com/daedan/festabook/presentation/home/HomeFragment.kt
@@ -1,10 +1,11 @@
-package com.daedan.festabook
+package com.daedan.festabook.presentation.home
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.daedan.festabook.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
